Enc common: EntropyCoding dependency removal
