+++ Source/Lib/Common/Codec/EbInterPrediction.h
@@ -13,6 +13,7 @@
 #include "EbMcp.h"
 #include "filter.h"
 #include "convolve.h"
+#include "EbCabacContextModel.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -28,6 +29,13 @@ extern "C" {
 #define RS_SCALE_SUBPEL_MASK ((1 << RS_SCALE_SUBPEL_BITS) - 1)
 #define RS_SCALE_EXTRA_BITS (RS_SCALE_SUBPEL_BITS - RS_SUBPEL_BITS)
 #define RS_SCALE_EXTRA_OFF (1 << (RS_SCALE_EXTRA_BITS - 1))
+#define MV_BORDER (16 << 3) // Allow 16 pels in 1/8th pel units
+
+
+#define NELEMENTS(x) (int)(sizeof(x) / sizeof(x[0]))
+
+#define INTRABC_DELAY_PIXELS 256 //  Delay of 256 pixels
+#define INTRABC_DELAY_SB64 (INTRABC_DELAY_PIXELS / 64)
 
     extern DECLARE_ALIGNED(256, const InterpKernel, sub_pel_filters_8[SUBPEL_SHIFTS]);
     extern DECLARE_ALIGNED(256, const InterpKernel, sub_pel_filters_4[SUBPEL_SHIFTS]);
@@ -192,6 +200,221 @@ int32_t get_wedge_bits_lookup(BlockSize sb_type);
 const uint8_t *av1_get_contiguous_soft_mask(int wedge_index, int wedge_sign, BlockSize sb_type);
 
 int get_wedge_params_bits(BlockSize sb_type);
+
+int is_masked_compound_type(COMPOUND_TYPE type);
+
+// Although we assign 32 bit integers, all the values are strictly under 14
+// bits.
+static int div_mult[32] = {0,    16384, 8192, 5461, 4096, 3276, 2730, 2340, 2048, 1820, 1638,
+                           1489, 1365,  1260, 1170, 1092, 1024, 963,  910,  862,  819,  780,
+                           744,  712,   682,  655,  630,  606,  585,  564,  546,  528};
+
+static INLINE void integer_mv_precision(MV *mv) {
+    int mod = (mv->row % 8);
+    if (mod != 0) {
+        mv->row -= mod;
+        if (abs(mod) > 4) {
+            if (mod > 0)
+                mv->row += 8;
+            else
+                mv->row -= 8;
+        }
+    }
+
+    mod = (mv->col % 8);
+    if (mod != 0) {
+        mv->col -= mod;
+        if (abs(mod) > 4) {
+            if (mod > 0)
+                mv->col += 8;
+            else
+                mv->col -= 8;
+        }
+    }
+}
+
+static INLINE void lower_mv_precision(MV *mv, int allow_hp, int is_integer) {
+    if (is_integer)
+        integer_mv_precision(mv);
+    else {
+        if (!allow_hp) {
+            if (mv->row & 1) mv->row += (mv->row > 0 ? -1 : 1);
+            if (mv->col & 1) mv->col += (mv->col > 0 ? -1 : 1);
+        }
+    }
+}
+
+static INLINE void get_mv_projection(MV *output, MV ref, int num, int den) {
+    den              = AOMMIN(den, MAX_FRAME_DISTANCE);
+    num              = num > 0 ? AOMMIN(num, MAX_FRAME_DISTANCE) : AOMMAX(num, -MAX_FRAME_DISTANCE);
+    const int mv_row = ROUND_POWER_OF_TWO_SIGNED(ref.row * num * div_mult[den], 14);
+    const int mv_col = ROUND_POWER_OF_TWO_SIGNED(ref.col * num * div_mult[den], 14);
+    const int clamp_max = MV_UPP - 1;
+    const int clamp_min = MV_LOW + 1;
+    output->row         = (int16_t)clamp(mv_row, clamp_min, clamp_max);
+    output->col         = (int16_t)clamp(mv_col, clamp_min, clamp_max);
+}
+
+static INLINE int check_sb_border(const int mi_row, const int mi_col, const int row_offset,
+                                  const int col_offset) {
+    const int sb_mi_size = mi_size_wide[BLOCK_64X64];
+    const int row        = mi_row & (sb_mi_size - 1);
+    const int col        = mi_col & (sb_mi_size - 1);
+
+    if (row + row_offset < 0 || row + row_offset >= sb_mi_size || col + col_offset < 0 ||
+        col + col_offset >= sb_mi_size)
+        return 0;
+
+    return 1;
+}
+
+static INLINE int is_neighbor_overlappable(const MbModeInfo *mbmi) {
+    return /*is_intrabc_block(mbmi) ||*/ mbmi->block_mi.ref_frame[0] >
+                                         INTRA_FRAME; // TODO: modify when add intra_bc
+}
+
+static INLINE int32_t is_mv_valid(const MV *mv) {
+    return mv->row > MV_LOW && mv->row < MV_UPP && mv->col > MV_LOW && mv->col < MV_UPP;
+}
+
+#define CHECK_BACKWARD_REFS(ref_frame) \
+    (((ref_frame) >= BWDREF_FRAME) && ((ref_frame) <= ALTREF_FRAME))
+#define IS_BACKWARD_REF_FRAME(ref_frame) CHECK_BACKWARD_REFS(ref_frame)
+
+void av1_find_ref_dv(IntMv *ref_dv, const TileInfo *const tile, int mib_size, int mi_row,
+                     int mi_col);
+
+static INLINE int32_t is_comp_ref_allowed(BlockSize bsize) {
+    return AOMMIN(block_size_wide[bsize], block_size_high[bsize]) >= 8;
+}
+
+static INLINE int is_interinter_compound_used(CompoundType type, BlockSize sb_type) {
+    const int comp_allowed = is_comp_ref_allowed(sb_type);
+    switch (type) {
+        case COMPOUND_AVERAGE:
+        case COMPOUND_DISTWTD:
+        case COMPOUND_DIFFWTD: return comp_allowed;
+        case COMPOUND_WEDGE: return comp_allowed && get_wedge_params_bits(sb_type) > 0;
+        default: assert(0); return 0;
+    }
+}
+
+static INLINE int is_any_masked_compound_used(BlockSize sb_type) {
+    CompoundType comp_type;
+    int          i;
+    if (!is_comp_ref_allowed(sb_type)) return 0;
+    for (i = 0; i < COMPOUND_TYPES; i++) {
+        comp_type = (CompoundType)i;
+        if (is_masked_compound_type(comp_type) && is_interinter_compound_used(comp_type, sb_type))
+            return 1;
+    }
+    return 0;
+}
+
+static INLINE int bsize_to_tx_size_cat(BlockSize bsize) {
+    TxSize tx_size = max_txsize_rect_lookup[bsize];
+    assert(tx_size != TX_4X4);
+    int depth = 0;
+    while (tx_size != TX_4X4) {
+        depth++;
+        tx_size = sub_tx_size_map[tx_size];
+        assert(depth < 10);
+    }
+    assert(depth <= MAX_TX_CATS);
+    return depth - 1;
+}
+
+static INLINE int bsize_to_max_depth(BlockSize bsize) {
+    TxSize tx_size = max_txsize_rect_lookup[bsize];
+    int    depth   = 0;
+    while (depth < MAX_TX_DEPTH && tx_size != TX_4X4) {
+        depth++;
+        tx_size = sub_tx_size_map[tx_size];
+    }
+    return depth;
+}
+
+static INLINE PredictionMode compound_ref0_mode(PredictionMode mode) {
+    static PredictionMode lut[] = {
+            MB_MODE_COUNT, // DC_PRED
+            MB_MODE_COUNT, // V_PRED
+            MB_MODE_COUNT, // H_PRED
+            MB_MODE_COUNT, // D45_PRED
+            MB_MODE_COUNT, // D135_PRED
+            MB_MODE_COUNT, // D113_PRED
+            MB_MODE_COUNT, // D157_PRED
+            MB_MODE_COUNT, // D203_PRED
+            MB_MODE_COUNT, // D67_PRED
+            MB_MODE_COUNT, // SMOOTH_PRED
+            MB_MODE_COUNT, // SMOOTH_V_PRED
+            MB_MODE_COUNT, // SMOOTH_H_PRED
+            MB_MODE_COUNT, // PAETH_PRED
+            MB_MODE_COUNT, // NEARESTMV
+            MB_MODE_COUNT, // NEARMV
+            MB_MODE_COUNT, // GLOBALMV
+            MB_MODE_COUNT, // NEWMV
+            NEARESTMV, // NEAREST_NEARESTMV
+            NEARMV, // NEAR_NEARMV
+            NEARESTMV, // NEAREST_NEWMV
+            NEWMV, // NEW_NEARESTMV
+            NEARMV, // NEAR_NEWMV
+            NEWMV, // NEW_NEARMV
+            GLOBALMV, // GLOBAL_GLOBALMV
+            NEWMV, // NEW_NEWMV
+    };
+    assert(NELEMENTS(lut) == MB_MODE_COUNT);
+    assert(is_inter_compound_mode(mode));
+    return lut[mode];
+}
+
+static INLINE PredictionMode compound_ref1_mode(PredictionMode mode) {
+    static PredictionMode lut[] = {
+            MB_MODE_COUNT, // DC_PRED
+            MB_MODE_COUNT, // V_PRED
+            MB_MODE_COUNT, // H_PRED
+            MB_MODE_COUNT, // D45_PRED
+            MB_MODE_COUNT, // D135_PRED
+            MB_MODE_COUNT, // D113_PRED
+            MB_MODE_COUNT, // D157_PRED
+            MB_MODE_COUNT, // D203_PRED
+            MB_MODE_COUNT, // D67_PRED
+            MB_MODE_COUNT, // SMOOTH_PRED
+            MB_MODE_COUNT, // SMOOTH_V_PRED
+            MB_MODE_COUNT, // SMOOTH_H_PRED
+            MB_MODE_COUNT, // PAETH_PRED
+            MB_MODE_COUNT, // NEARESTMV
+            MB_MODE_COUNT, // NEARMV
+            MB_MODE_COUNT, // GLOBALMV
+            MB_MODE_COUNT, // NEWMV
+            NEARESTMV, // NEAREST_NEARESTMV
+            NEARMV, // NEAR_NEARMV
+            NEWMV, // NEAREST_NEWMV
+            NEARESTMV, // NEW_NEARESTMV
+            NEWMV, // NEAR_NEWMV
+            NEARMV, // NEW_NEARMV
+            GLOBALMV, // GLOBAL_GLOBALMV
+            NEWMV, // NEW_NEWMV
+    };
+    assert(NELEMENTS(lut) == MB_MODE_COUNT);
+    assert(is_inter_compound_mode(mode));
+    return lut[mode];
+}
+
+static INLINE EbBool is_motion_variation_allowed_bsize(const BlockSize bsize) {
+    return (block_size_wide[bsize] >= 8 && block_size_high[bsize] >= 8);
+}
+
+static INLINE int is_global_mv_block(const PredictionMode mode, const BlockSize bsize,
+                                     TransformationType type) {
+    return (mode == GLOBALMV || mode == GLOBAL_GLOBALMV) && type > TRANSLATION &&
+           is_motion_variation_allowed_bsize(bsize);
+}
+
+static INLINE uint32_t have_nearmv_in_inter_mode(PredictionMode mode) {
+    return (mode == NEARMV || mode == NEAR_NEARMV || mode == NEAR_NEWMV || mode == NEW_NEARMV);
+}
+
+static INLINE int is_intrabc_block(const BlockModeInfo *block_mi) { return block_mi->use_intrabc; }
 
 
 #ifdef __cplusplus
