+++ Source/Lib/Decoder/Codec/EbDecCdef.h
@@ -19,6 +19,7 @@
 
 #include "EbDecHandle.h"
 #include "EbDecUtils.h"
+#include "../../Encoder/Codec/EbCdef.h"
 
 #ifdef __cplusplus
 extern "C" {
