+++ Source/Lib/Decoder/Codec/EbDecInverseQuantize.c
@@ -24,6 +24,7 @@
 #include "EbDecProcessFrame.h"
 #include "EbCoefficients.h"
 #include "EbQMatrices.h"
+#include "EbInvTransforms.h"
 
 // Same wrapper(av1_ac/dc_quant_qtx) available in .c file of encoder
 static INLINE int16_t get_dc_quant(int32_t qindex, int32_t delta, AomBitDepth bit_depth) {
