+++ Source/Lib/Decoder/Codec/EbDecParseBlock.c
@@ -31,9 +31,11 @@
 
 #include "EbDecParseInterBlock.h"
 #include "EbDecProcessFrame.h"
-#include "../../Encoder/Codec/EbEntropyCoding.h"
+
 #include "EbCommonUtils.h"
 #include "EbCoefficients.h"
+#include "EbIntraPrediction.h"
+#include "EbCabacContextModel.h"
 
 #include "EbLog.h"
